Move baskets() to PersonaBO
